Система координат сервера сделана одномерной (ТЗ D.1)
Для наглядности система координат клиента также изменена на одномерную, логика парсинга аргументов командной строки изменена, так как используемая библиотека не поддерживает парсинг массива с 1 элементом. Дополнительно добавлен шаблонный параметр S в Location для тестирования различных систем координат в common/tests.

+++ server/app/menu/menu_item/menu_item_dist/menu_item_dist_impl.h
@@ -12,7 +12,7 @@ template <typename T> std::string_view MenuItemDist<T>::getName() const {
   return "move";
 }
 
-template <typename T> int MenuItemDist<T>::getArgsCount() {
+template <typename T> size_t MenuItemDist<T>::getArgsCount() {
   return common::constants::LOCATION_COORDS_COUNT;
 }
